Initial Commit
This first commit marks me finding a suitable parser (or so it seems) for me to
develop the syntax checker with. For syntax rules, the Gendarme libraries should
prove an instructive resource.

A project build system has been set up, and the parser compiles and runs (using
itself as test data).

---
M.Edwards